Fixed exponent and modulous
